refactor post processing, render settings and media settings
