+++ packages/editor/src/functions/PrefabEditors.tsx
@@ -24,7 +24,6 @@ import { OceanComponent } from '@xrengine/engine/src/scene/components/OceanCompo
 import { ParticleEmitterComponent } from '@xrengine/engine/src/scene/components/ParticleEmitterComponent'
 import { PointLightComponent } from '@xrengine/engine/src/scene/components/PointLightComponent'
 import { PortalComponent } from '@xrengine/engine/src/scene/components/PortalComponent'
-import { PostprocessingComponent } from '@xrengine/engine/src/scene/components/PostprocessingComponent'
 import { ScenePreviewCameraComponent } from '@xrengine/engine/src/scene/components/ScenePreviewCamera'
 import { SceneTagComponent } from '@xrengine/engine/src/scene/components/SceneTagComponent'
 import { SkyboxComponent } from '@xrengine/engine/src/scene/components/SkyboxComponent'
@@ -78,7 +77,7 @@ import VideoNodeEditor from '../components/properties/VideoNodeEditor'
 import VolumetricNodeEditor from '../components/properties/VolumetricNodeEditor'
 import WaterNodeEditor from '../components/properties/WaterNodeEditor'
 
-export const EntityNodeEditor = new Map<Component<any, any, any>, EditorComponentType>()
+export const EntityNodeEditor = new Map<Component, EditorComponentType>()
 EntityNodeEditor.set(TransformComponent, TransformPropertyGroup)
 EntityNodeEditor.set(DirectionalLightComponent, DirectionalLightNodeEditor)
 EntityNodeEditor.set(HemisphereLightComponent, HemisphereLightNodeEditor)
@@ -94,7 +93,6 @@ EntityNodeEditor.set(MountPointComponent, MountPointNodeEditor)
 EntityNodeEditor.set(ColliderComponent, ColliderNodeEditor)
 EntityNodeEditor.set(GroupComponent, GroupNodeEditor)
 EntityNodeEditor.set(AssetComponent, AssetNodeEditor)
-EntityNodeEditor.set(PostprocessingComponent, PostProcessingNodeEditor)
 EntityNodeEditor.set(SceneTagComponent, SceneNodeEditor)
 EntityNodeEditor.set(ScenePreviewCameraComponent, ScenePreviewCameraNodeEditor)
 EntityNodeEditor.set(SkyboxComponent, SkyboxNodeEditor)
