--- packages/engine/src/audio/components/ImmersiveMediaTagComponent.ts
@@ -1,5 +1,0 @@
-import { createMappedComponent } from '../../ecs/functions/ComponentFunctions'
-
-export const ImmersiveMediaTagComponent = createMappedComponent<true>('ImmersiveMediaTagComponent')
-
-export const SCENE_COMPONENT_MEDIA_SETTINGS = 'media-settings'
