+++ packages/engine/src/ecs/classes/Engine.ts
@@ -2,10 +2,8 @@ import type { UserId } from '@xrengine/common/src/interfaces/UserId'
 import { createHyperStore, getState } from '@xrengine/hyperflux'
 import { HyperStore } from '@xrengine/hyperflux/functions/StoreFunctions'
 
-import { SCENE_COMPONENT_AUDIO_SETTINGS_DEFAULT_VALUES } from '../../audio/components/PositionalAudioSettingsComponent'
 import { NetworkTopics } from '../../networking/classes/Network'
 import type { World } from '../classes/World'
-import type { SystemModuleType } from '../functions/SystemFunctions'
 
 import '../utils/threejsPatches'
 
@@ -40,11 +38,6 @@ export class Engine {
     notifications: null! as GainNode,
     music: null! as GainNode,
     soundEffects: null! as GainNode
-  }
-
-  /** @todo refactor this into hyperflux state */
-  spatialAudioSettings = {
-    ...SCENE_COMPONENT_AUDIO_SETTINGS_DEFAULT_VALUES
   }
 
   /**
