+++ packages/engine/src/ecs/functions/EntityTree.ts
@@ -4,6 +4,7 @@ import { getState } from '@xrengine/hyperflux'
 
 import { NameComponent } from '../../scene/components/NameComponent'
 import { SceneObjectComponent } from '../../scene/components/SceneObjectComponent'
+import { SceneTagComponent } from '../../scene/components/SceneTagComponent'
 import { VisibleComponent } from '../../scene/components/VisibleComponent'
 import { serializeEntity } from '../../scene/functions/serializeWorld'
 import {
@@ -15,7 +16,7 @@ import { updateEntityTransform } from '../../transform/systems/TransformSystem'
 import { Engine } from '../classes/Engine'
 import { EngineState } from '../classes/EngineState'
 import { Entity } from '../classes/Entity'
-import { addComponent, getComponent, hasComponent } from '../functions/ComponentFunctions'
+import { addComponent, getComponent, hasComponent, setComponent } from '../functions/ComponentFunctions'
 import { createEntity, entityExists, removeEntity } from '../functions/EntityFunctions'
 
 // Data structure to hold parent child relationship between entities
@@ -61,8 +62,9 @@ export function initializeEntityTree(world = Engine.instance.currentWorld): void
   if (entityExists(world.sceneEntity)) removeEntity(world.sceneEntity, true)
 
   world.sceneEntity = createEntity()
-  addComponent(world.sceneEntity, NameComponent, { name: 'scene' })
-  addComponent(world.sceneEntity, VisibleComponent, true)
+  setComponent(world.sceneEntity, NameComponent, { name: 'scene' })
+  setComponent(world.sceneEntity, VisibleComponent, true)
+  setComponent(world.sceneEntity, SceneTagComponent, true)
   setTransformComponent(world.sceneEntity)
   // addObjectToGroup(world.sceneEntity, world.scene)
