+++ packages/engine/src/scene/components/SceneTagComponent.ts
@@ -2,4 +2,4 @@ import { createMappedComponent } from '../../ecs/functions/ComponentFunctions'
 
 export const SCENE_COMPONENT_SCENE_TAG = 'scene-tag'
 
-export const SceneTagComponent = createMappedComponent('SceneTagComponent')
+export const SceneTagComponent = createMappedComponent<true>('SceneTagComponent')
