+++ packages/engine/src/scene/systems/SceneObjectUpdateSystem.ts
@@ -15,7 +15,6 @@ import { Engine } from '../../ecs/classes/Engine'
 import { EngineActions } from '../../ecs/classes/EngineState'
 import { World } from '../../ecs/classes/World'
 import { defineQuery, getComponent, hasComponent, removeQuery } from '../../ecs/functions/ComponentFunctions'
-import { configureEffectComposer } from '../../renderer/functions/configureEffectComposer'
 import {
   SCENE_COMPONENT_TRANSFORM,
   SCENE_COMPONENT_TRANSFORM_DEFAULT_VALUES
@@ -70,17 +69,7 @@ import {
   SCENE_COMPONENT_PORTAL_COLLIDER_VALUES,
   SCENE_COMPONENT_PORTAL_DEFAULT_VALUES
 } from '../components/PortalComponent'
-import {
-  PostprocessingComponent,
-  SCENE_COMPONENT_POSTPROCESSING,
-  SCENE_COMPONENT_POSTPROCESSING_DEFAULT_VALUES
-} from '../components/PostprocessingComponent'
 import { PreventBakeTagComponent, SCENE_COMPONENT_PREVENT_BAKE } from '../components/PreventBakeTagComponent'
-import {
-  RenderSettingComponent,
-  SCENE_COMPONENT_RENDERER_SETTINGS,
-  SCENE_COMPONENT_RENDERER_SETTINGS_DEFAULT_VALUES
-} from '../components/RenderSettingComponent'
 import { SceneAssetPendingTagComponent } from '../components/SceneAssetPendingTagComponent'
 import { SCENE_COMPONENT_SCENE_PREVIEW_CAMERA, ScenePreviewCameraComponent } from '../components/ScenePreviewCamera'
 import { SceneTagComponent } from '../components/SceneTagComponent'
@@ -139,17 +128,6 @@ import { deserializeModel, serializeModel } from '../functions/loaders/ModelFunc
 import { deserializeOcean, serializeOcean, updateOcean } from '../functions/loaders/OceanFunctions'
 import { deserializePortal, serializePortal, updatePortal } from '../functions/loaders/PortalFunctions'
 import {
-  deserializePostprocessing,
-  serializePostprocessing,
-  shouldDeserializePostprocessing
-} from '../functions/loaders/PostprocessingFunctions'
-import {
-  deserializeRenderSetting,
-  resetEngineRenderer,
-  serializeRenderSettings,
-  updateRenderSetting
-} from '../functions/loaders/RenderSettingsFunction'
-import {
   enterScenePreviewCamera,
   serializeScenePreviewCamera,
   shouldDeserializeScenePreviewCamera
@@ -177,7 +155,6 @@ export const ScenePrefabs = {
   particleEmitter: 'Particle Emitter' as const,
   portal: 'Portal' as const,
   chair: 'Chair' as const,
-  postProcessing: 'Post Processing' as const,
   previewCamera: 'Preview Camera' as const,
   skybox: 'Skybox' as const,
   spawnPoint: 'Spawn Point' as const,
@@ -255,26 +232,6 @@ export default async function SceneObjectUpdateSystem(world: World) {
   world.sceneComponentRegistry.set(SpawnPointComponent.name, SCENE_COMPONENT_SPAWN_POINT)
   world.sceneLoadingRegistry.set(SCENE_COMPONENT_SPAWN_POINT, {
     defaultData: SCENE_COMPONENT_SPAWN_POINT_DEFAULT_DATA
-  })
-
-  world.sceneComponentRegistry.set(RenderSettingComponent.name, SCENE_COMPONENT_RENDERER_SETTINGS)
-  world.sceneLoadingRegistry.set(SCENE_COMPONENT_RENDERER_SETTINGS, {
-    defaultData: SCENE_COMPONENT_RENDERER_SETTINGS_DEFAULT_VALUES,
-    deserialize: deserializeRenderSetting,
-    serialize: serializeRenderSettings
-  })
-
-  world.scenePrefabRegistry.set(ScenePrefabs.postProcessing, [
-    { name: SCENE_COMPONENT_VISIBLE, props: true },
-    { name: SCENE_COMPONENT_POSTPROCESSING, props: SCENE_COMPONENT_POSTPROCESSING_DEFAULT_VALUES }
-  ])
-
-  world.sceneComponentRegistry.set(PostprocessingComponent.name, SCENE_COMPONENT_POSTPROCESSING)
-  world.sceneLoadingRegistry.set(SCENE_COMPONENT_POSTPROCESSING, {
-    defaultData: SCENE_COMPONENT_POSTPROCESSING_DEFAULT_VALUES,
-    deserialize: deserializePostprocessing,
-    serialize: serializePostprocessing,
-    shouldDeserialize: shouldDeserializePostprocessing
   })
 
   /**
@@ -497,8 +454,6 @@ export default async function SceneObjectUpdateSystem(world: World) {
   const cloudQuery = defineQuery([CloudComponent])
   const oceanQuery = defineQuery([OceanComponent])
   const interiorQuery = defineQuery([InteriorComponent])
-  const renderSettingsQuery = defineQuery([RenderSettingComponent])
-  const postProcessingQuery = defineQuery([PostprocessingComponent])
   const cameraPropertiesQuery = defineQuery([CameraPropertiesComponent, FollowCameraComponent, CameraComponent])
   const scenePreviewCameraQuery = defineQuery([ScenePreviewCameraComponent])
 
@@ -520,8 +475,6 @@ export default async function SceneObjectUpdateSystem(world: World) {
         if (hasComponent(entity, CloudComponent)) updateCloud(entity)
         if (hasComponent(entity, OceanComponent)) updateOcean(entity)
         if (hasComponent(entity, InteriorComponent)) updateInterior(entity)
-        if (hasComponent(entity, RenderSettingComponent)) updateRenderSetting(entity)
-        if (hasComponent(entity, PostprocessingComponent)) configureEffectComposer()
         if (hasComponent(entity, CameraPropertiesComponent)) updateCameraProperties(entity)
       }
     }
@@ -552,10 +505,6 @@ export default async function SceneObjectUpdateSystem(world: World) {
     for (const entity of cloudQuery.enter()) updateCloud(entity)
     for (const entity of oceanQuery.enter()) updateOcean(entity)
     for (const entity of interiorQuery.enter()) updateInterior(entity)
-    for (const entity of renderSettingsQuery.enter()) updateRenderSetting(entity)
-    for (const entity of renderSettingsQuery.exit()) if (!renderSettingsQuery().length) resetEngineRenderer(true)
-    for (const entity of postProcessingQuery.enter()) configureEffectComposer()
-    for (const entity of postProcessingQuery.exit()) configureEffectComposer()
     for (const entity of cameraPropertiesQuery.enter()) updateCameraProperties(entity)
     for (const entity of scenePreviewCameraQuery.enter()) enterScenePreviewCamera(entity)
   }
@@ -593,14 +542,6 @@ export default async function SceneObjectUpdateSystem(world: World) {
 
     world.sceneComponentRegistry.delete(SpawnPointComponent.name)
     world.sceneLoadingRegistry.delete(SCENE_COMPONENT_SPAWN_POINT)
-
-    world.sceneComponentRegistry.delete(RenderSettingComponent.name)
-    world.sceneLoadingRegistry.delete(SCENE_COMPONENT_RENDERER_SETTINGS)
-
-    world.scenePrefabRegistry.delete(ScenePrefabs.postProcessing)
-
-    world.sceneComponentRegistry.delete(PostprocessingComponent.name)
-    world.sceneLoadingRegistry.delete(SCENE_COMPONENT_POSTPROCESSING)
 
     /**
      * Assets
@@ -710,8 +651,6 @@ export default async function SceneObjectUpdateSystem(world: World) {
     removeQuery(world, cloudQuery)
     removeQuery(world, oceanQuery)
     removeQuery(world, interiorQuery)
-    removeQuery(world, renderSettingsQuery)
-    removeQuery(world, postProcessingQuery)
     removeQuery(world, cameraPropertiesQuery)
     removeQuery(world, scenePreviewCameraQuery)
